quick setup dialog now deals with entire modpack zips complete with maplist
