+++ quicksetupdialog.ui
@@ -10,7 +10,7 @@
     <x>0</x>
     <y>0</y>
     <width>919</width>
-    <height>254</height>
+    <height>242</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -20,7 +20,7 @@
    <bool>true</bool>
   </property>
   <layout class="QGridLayout" name="gridLayout">
-   <item row="4" column="2" colspan="2">
+   <item row="3" column="2" colspan="2">
     <widget class="QCheckBox" name="separateSaveGame">
      <property name="text">
       <string>Separate Save Game</string>
@@ -30,25 +30,18 @@
    <item row="1" column="0">
     <widget class="QLabel" name="label_2">
      <property name="text">
-      <string>Modlist file</string>
+      <string>Modpack .zip</string>
      </property>
     </widget>
    </item>
-   <item row="2" column="2" colspan="2">
-    <widget class="QPushButton" name="chooseMapListFile">
-     <property name="text">
-      <string>Choose Map List File</string>
-     </property>
-    </widget>
-   </item>
-   <item row="3" column="1">
+   <item row="2" column="1">
     <widget class="QLineEdit" name="riivolutionPatchName">
      <property name="text">
       <string>csmm</string>
      </property>
     </widget>
    </item>
-   <item row="3" column="2" colspan="2">
+   <item row="2" column="2" colspan="2">
     <widget class="QCheckBox" name="enableRiivolution">
      <property name="text">
       <string>Enable Riivolution</string>
@@ -62,14 +55,14 @@
      </property>
     </widget>
    </item>
-   <item row="4" column="0">
+   <item row="3" column="0">
     <widget class="QLabel" name="label_5">
      <property name="text">
       <string>Marker Code</string>
      </property>
     </widget>
    </item>
-   <item row="4" column="1">
+   <item row="3" column="1">
     <widget class="QLineEdit" name="markerCode"/>
    </item>
    <item row="0" column="1">
@@ -86,28 +79,21 @@
      </property>
     </widget>
    </item>
-   <item row="3" column="0">
+   <item row="2" column="0">
     <widget class="QLabel" name="riivolutionPatchNameLabel">
      <property name="text">
       <string>Riivolution Patch Name</string>
      </property>
     </widget>
    </item>
-   <item row="2" column="1">
-    <widget class="QLineEdit" name="mapListFile">
-     <property name="enabled">
-      <bool>false</bool>
-     </property>
-    </widget>
-   </item>
    <item row="1" column="2" colspan="2">
-    <widget class="QPushButton" name="chooseModpackFile">
+    <widget class="QPushButton" name="chooseModpackZip">
      <property name="text">
-      <string>Choose Mod List File</string>
+      <string>Choose Modpack .zip</string>
      </property>
     </widget>
    </item>
-   <item row="5" column="0" colspan="4">
+   <item row="4" column="0" colspan="4">
     <widget class="QDialogButtonBox" name="buttonBox">
      <property name="enabled">
       <bool>true</bool>
@@ -127,15 +113,8 @@
      </property>
     </widget>
    </item>
-   <item row="2" column="0">
-    <widget class="QLabel" name="label_3">
-     <property name="text">
-      <string>Maplist file</string>
-     </property>
-    </widget>
-   </item>
    <item row="1" column="1">
-    <widget class="QLineEdit" name="modpackFile">
+    <widget class="QLineEdit" name="modpackZip">
      <property name="enabled">
       <bool>false</bool>
      </property>
